+++ packages/i18n/src/t/index.ts
@@ -1,4 +1,4 @@
 export * from './translation-renderer';
 export * from './translation-parameters-renderer';
 export * from './translation-binding';
-export * from './value-converter';
+export * from './translation-value-converter';
